[Data Object] Objectbricks shouldn't extend from Field Collections #5774
